Split the clipping logic into a separate class
